Extract query output format helpers
